made email validation, added confirm delete for user and role, change check on super admin, fixed bugs

+++ src/app/features/role/pages/role-list/role-list.component.ts
@@ -1,5 +1,4 @@
 import { AfterViewInit, Component, ViewChild } from '@angular/core';
-
 import { MatPaginator } from '@angular/material/paginator';
 import { MatTableDataSource } from '@angular/material/table';
 import { MatSort, Sort } from '@angular/material/sort';
@@ -13,6 +12,8 @@ import { RoleCategoryFirebaseService } from '@core/services/firebase/firebase-en
 import { CategoryFirebaseService } from '@core/services/firebase/firebase-entities/categoryFirebase.service';
 import { IRoleCategoryPair } from '@core/models/RoleCategoryPair';
 import { ICategory } from '@core/models/Category';
+import { ConfirmationSheetChoice } from '@shared/components/confirmation-dialog/confirmation-dialog.component';
+import { ConfirmationDialogService } from '@core/services/confirmationDialog/confirmationDialog.service';
 
 @Component({
 	selector: 'app-user',
@@ -36,7 +37,8 @@ export class RoleListComponent implements AfterViewInit {
 		private router: Router,
 		private roleFirebaseService: RoleFirebaseService,
 		private roleCategoryFirebaseService: RoleCategoryFirebaseService,
-		private categoryFirebaseService: CategoryFirebaseService
+		private categoryFirebaseService: CategoryFirebaseService,
+		private confirmationDialogService: ConfirmationDialogService
 	) {}
 
 	@ViewChild(MatPaginator) paginator: MatPaginator;
@@ -80,18 +82,37 @@ export class RoleListComponent implements AfterViewInit {
 		}
 	}
 
+	confirmationDelete(role: IRole) {
+		this.confirmationDialogService
+			.ask(`Are you sure to delete role ${role.name}?`)
+			.then(result => {
+				if (result === ConfirmationSheetChoice.CONFIRMED) {
+					this.onDelete(role);
+				}
+			});
+	}
+
 	onDelete(role: IRole): void {
 		if (role.id !== '0') {
-			this.roleFirebaseService.deleteRole(role.id);
-			role.availableCategoriesToView.map(item => {
-				this.roleCategoryFirebaseService
-					.getRoleCategoriesId(role.id, item.id)
-					.pipe(take(1))
-					.subscribe((res: IRoleCategoryPair[]) => {
-						res.map(item => {
-							this.roleCategoryFirebaseService.deleteDoc(item.id);
-						});
+			this.roleFirebaseService.deleteRole(role.id).then(() => {
+				if (
+					role.availableCategoriesToView &&
+					role.availableCategoriesToView.length
+				) {
+					role.availableCategoriesToView.map(item => {
+						this.roleCategoryFirebaseService
+							.getRoleCategoriesId(role.id, item.id)
+							.pipe(take(1))
+							.subscribe((res: IRoleCategoryPair[]) => {
+								res.map(item => {
+									this.roleCategoryFirebaseService.deleteDoc(item.id);
+								});
+							});
 					});
+				}
+				this.snackBService.openSnackBar(
+					`The ${role.name} role has deleted successfully`
+				);
 			});
 		} else {
 			this.snackBService.openSnackBar('This Super Admin role cannot delete');
